Move common types to shared/types.ts and remove duplicate types.ts file
