--- pages/src/types.ts
@@ -1,9 +1,0 @@
-export interface HealthStatus {
-  healthy: boolean;
-  error?: string;
-}
-
-export interface Location {
-  hostname: string;
-  [key: string]: unknown;
-}
