+++ shared/types.ts
@@ -11,3 +11,13 @@ export interface PageContent {
   summary: string;
   extractedAt: number;
 }
+
+export interface HealthStatus {
+  healthy: boolean;
+  error?: string;
+}
+
+export interface Location {
+  hostname: string;
+  [key: string]: unknown;
+}
